chore(tracing): Split tracing utils into separate module (#952)
No behavior changes - just moving stuff around.
# Conflicts:
#	sentry_sdk/tracing.py
